fix: Item dto 수정 #51

+++ monew/src/main/java/com/codeit/team2/monew/module/domain/useractivity/dto/ArticleViewItem.java
@@ -13,8 +13,8 @@ public record ArticleViewItem(
     String articleTitle,
     Instant articlePublishedDate,
     String articleSummary,
-    Integer articleCommentCount,
-    Integer articleViewCount
+    Long articleCommentCount,
+    Long articleViewCount
 ) {
 
 }
