+++ monew/src/main/java/com/codeit/team2/monew/module/domain/useractivity/dto/SubscriptionItem.java
@@ -9,7 +9,7 @@ public record SubscriptionItem(
     UUID interestId,
     String interestName,
     List<String> interestKeywords,
-    Integer interestSubscriberCount,
+    Long interestSubscriberCount,
     Instant createdAt
 ) {
